refactor: centralize shared TUI constants into ui.ts
Rename useTerminalWidth.ts → ui.ts and add shared display constants:
- CURSOR ('▊') — standardized across all text input fields
- MONTHS — was duplicated in 4 files
- FLEX_COLORS — was duplicated in 3 files with inconsistent types
- SUBTYPE_DISPLAY — account subtype aliases (e.g. 'crypto exchange' → 'crypto')
